feat: add POST /api/events/:id/stats endpoint for batch stats import
Extract shared logic into MatchStatsImportable concern and expose a
new API action that accepts a JSON array from the scraper workflow,
matching matches by position order (same approach as timestamps API).
