+++ app/controllers/api/events_controller.rb
@@ -1,6 +1,7 @@
 module Api
   class EventsController < BaseController
     include TimestampParseable
+    include MatchStatsImportable
 
     def timestamps
       event = Event.find_by(id: params[:id])
@@ -45,6 +46,37 @@ module Api
       error_msg = params[:error].presence || "不明なエラー"
       PushNotificationService.notify_timestamps_failed(event: event, error: error_msg)
       render json: { message: "OK" }
+    end
+
+    # POST /api/events/:id/stats
+    def stats
+      event = Event.find_by(id: params[:id])
+      return render json: { error: "Event not found" }, status: :not_found unless event
+
+      matches_data = request.request_parameters
+      unless matches_data.is_a?(Array)
+        return render json: { error: "リクエストボディは JSON 配列である必要があります" }, status: :unprocessable_entity
+      end
+
+      db_matches = event.matches.includes(:match_timeline, match_players: [ :user ]).order(:played_at, :id)
+
+      if matches_data.size != db_matches.size
+        return render json: {
+          error: "データ件数（#{matches_data.size}）とイベント内試合数（#{db_matches.size}）が一致しません"
+        }, status: :unprocessable_entity
+      end
+
+      ActiveRecord::Base.transaction do
+        db_matches.each_with_index do |match, i|
+          @match = match
+          apply_timeline_data(matches_data[i])
+          recalculate_match_ranks
+        end
+      end
+
+      render json: { message: "OK", updated: db_matches.size }
+    rescue ActiveRecord::RecordInvalid => e
+      render json: { error: e.message }, status: :unprocessable_entity
     end
   end
 end
